+++ virvo/virvo/vvsoftrayrend.cpp
@@ -313,7 +313,7 @@ void vvSoftRayRend::renderTile(const vvSoftRayRend::Tile& tile, const vvMatrix& 
 
   vvsize3 minVox = _visibleRegion.getMin();
   vvsize3 maxVox = _visibleRegion.getMax();
-  for (int i = 0; i < 3; ++i)
+  for (size_t i = 0; i < 3; ++i)
   {
     minVox[i] = std::max(minVox[i], size_t(0));
     maxVox[i] = std::min(maxVox[i], vd->vox[i]);
@@ -401,7 +401,7 @@ void vvSoftRayRend::renderTile(const vvSoftRayRend::Tile& tile, const vvMatrix& 
           pos += step;
         }
 
-        for (int c = 0; c < 4; ++c)
+        for (size_t c = 0; c < 4; ++c)
         {
           (*colors)[y * _width * 4 + x * 4 + c] = dst[c];
         }
